+++ compiler-tests/src/test/data/diagnostic/inject/assisted/LazyAssistedFactoryInjection.kt
@@ -1,0 +1,9 @@
+// RENDER_DIAGNOSTICS_FULL_TEXT
+
+@Inject
+class Foo(factoryProvider: <!ASSISTED_FACTORIES_CANNOT_BE_LAZY!>Lazy<Factory><!>) {
+  @AssistedFactory
+  interface Factory {
+    fun create(): Foo
+  }
+}