+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/InjectConstructorChecker.kt
@@ -5,8 +5,10 @@ package dev.zacsweers.metro.compiler.fir.checkers
 import dev.zacsweers.metro.compiler.Symbols.DaggerSymbols
 import dev.zacsweers.metro.compiler.fir.FirMetroErrors
 import dev.zacsweers.metro.compiler.fir.annotationsIn
+import dev.zacsweers.metro.compiler.fir.asFirContextualTypeKey
 import dev.zacsweers.metro.compiler.fir.classIds
 import dev.zacsweers.metro.compiler.fir.findInjectConstructor
+import dev.zacsweers.metro.compiler.fir.isAnnotatedWithAny
 import dev.zacsweers.metro.compiler.fir.validateInjectedClass
 import dev.zacsweers.metro.compiler.fir.validateVisibility
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
@@ -17,6 +19,8 @@ import org.jetbrains.kotlin.fir.analysis.checkers.declaration.FirClassChecker
 import org.jetbrains.kotlin.fir.declarations.FirClass
 import org.jetbrains.kotlin.fir.declarations.getAnnotationByClassId
 import org.jetbrains.kotlin.fir.declarations.primaryConstructorIfAny
+import org.jetbrains.kotlin.fir.resolve.toClassSymbol
+import org.jetbrains.kotlin.fir.types.classId
 
 internal object InjectConstructorChecker : FirClassChecker(MppCheckerKind.Common) {
   override fun check(declaration: FirClass, context: CheckerContext, reporter: DiagnosticReporter) {
@@ -55,9 +59,30 @@ internal object InjectConstructorChecker : FirClassChecker(MppCheckerKind.Common
       return
     }
 
-    val constructorToValidate = injectedConstructor ?: declaration.primaryConstructorIfAny(session)
-    constructorToValidate?.validateVisibility(context, reporter, "Injected constructors") {
+    val constructorToValidate =
+      injectedConstructor ?: declaration.primaryConstructorIfAny(session) ?: return
+    constructorToValidate.validateVisibility(context, reporter, "Injected constructors") {
       return
+    }
+
+    for (parameter in constructorToValidate.valueParameterSymbols) {
+      if (parameter.isAnnotatedWithAny(session, classIds.assistedAnnotations)) continue
+      val type = parameter.resolvedReturnTypeRef.coneType
+      val contextKey = type.asFirContextualTypeKey(session, null, false)
+      if (contextKey.isWrappedInLazy) {
+        val canonicalType = contextKey.typeKey.type
+        val canonicalClass = canonicalType.toClassSymbol(session) ?: continue
+        if (canonicalClass.isAnnotatedWithAny(session, classIds.assistedFactoryAnnotations)) {
+          reporter.reportOn(
+            parameter.resolvedReturnTypeRef.source ?: parameter.source ?: source,
+            FirMetroErrors.ASSISTED_FACTORIES_CANNOT_BE_LAZY,
+            canonicalClass.name.asString(),
+            canonicalClass.classId.asFqNameString(),
+            context,
+          )
+          return
+        }
+      }
     }
   }
 }