+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/graph/BaseBinding.kt
@@ -20,5 +20,8 @@ internal interface BaseBinding<
   val isTransient: Boolean
     get() = false
 
+  val isImplicitlyDeferrable: Boolean
+    get() = contextualTypeKey.isDeferrable
+
   fun renderLocationDiagnostic(): String
 }