+++ CHANGELOG.md
@@ -5,7 +5,10 @@ Changelog
 --------------
 
 - **Enhancement:** Disambiguate `MetroContribution` class names based on scope to better support IC when changing scopes.
+- **Enhancement:** Minimize deferred types when breaking cycles.
+- **Fix:** Disallow injection of `Lazy<T>` where `T` is an `@AssistedFactory`-annotated class.
 - **Fix:** Don't short-circuit assisted injection validation if only an accessor exists.
+- **Fix:** Allow cycles of assisted factories to their target classes.
 
 0.3.7
 -----